+++ src/frontend/components/UI/QuestsViewer/components/QuestDetailsWrapper/index.module.scss
@@ -1,0 +1,6 @@
+.questDetails {
+  flex-grow: 1;
+  width: 100%;
+  max-width: 576px;
+  padding: var(--space-2lg);
+}
