+++ src/frontend/components/UI/QuestsViewer/components/QuestLogWrapper/index.module.scss
@@ -1,0 +1,4 @@
+.questLog {
+  flex-grow: 1;
+  margin: 0px var(--space-sm);
+}
