+++ src/frontend/components/UI/QuestsViewer/index.module.scss
@@ -7,16 +7,4 @@
   z-index: 4;
   display: flex;
   flex-direction: row;
-
-  > *.questLog {
-    flex-grow: 1;
-    margin: 0px var(--space-sm);
-  }
-
-  > *.questDetails {
-    flex-grow: 1;
-    width: 100%;
-    max-width: 576px;
-    padding: var(--space-2lg);
-  }
 }
